pyproject.toml
[project]
name = "fastapi-agent"
version = "0.1.0"
description = "FastAPI agent with Ollama/LiteLLM and MCP tools"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.95",
    "uvicorn[standard]>=0.22",
    "llama-index>=0.12.0",
    "llama-index-llms-ollama",
    "llama-index-llms-litellm",
    "llama-index-tools-mcp",
    "python-dotenv>=1.0",
    "httpx>=0.24.0",
]